+++ responder/__init__.py
@@ -7,10 +7,11 @@ including the API, Request, and Response classes.
 
 from . import ext
 from .__version__ import __version__
-from .core import API, Request, Response, abort
+from .core import API, HTTPMethod, Request, Response, abort
 
 __all__ = [
     "API",
+    "HTTPMethod",
     "Request",
     "Response",
     "__version__",