+++ responder/core.py
@@ -1,8 +1,9 @@
 from .api import API, abort
-from .models import Request, Response
+from .models import HTTPMethod, Request, Response
 
 __all__ = [
     "API",
+    "HTTPMethod",
     "Request",
     "Response",
     "abort",